Data.GADT.Show

Documentation

>>> import Data.Some

class GShow t where

Show-like class for 1-type-parameter GADTs. GShow t => ... is equivalent to something like (forall a. Show (t a)) => .... The easiest way to create instances would probably be to write (or derive) an instance Show (T a), and then simply say:

instance GShow t where gshowsPrec = showsPrec

Methods

gshowsPrec :: Int -> t a -> ShowS

gshows :: GShow t => t a -> ShowS

gshow :: GShow t => t a -> String

newtype GReadResult t

Constructors

GReadResult

Fields

getGReadResult :: forall b. (forall a. t a -> b) -> b

type GReadS t = String -> [(GReadResult t, String)]

GReadS t is equivalent to ReadS (forall b. (forall a. t a -> b) -> b), which is in turn equivalent to ReadS (Exists t) (with data Exists t where Exists :: t a -> Exists t)

class GRead t where

Read-like class for 1-type-parameter GADTs. Unlike GShow, this one cannot be mechanically derived from a Read instance because greadsPrec must choose the phantom type based on the String being parsed.

Methods

greadsPrec :: Int -> GReadS t

greads :: GRead t => GReadS t

gread :: GRead t => String -> (forall a. t a -> b) -> b

greadMaybe :: GRead t => String -> (forall a. t a -> b) -> Maybe b

>>> greadMaybe "InL Refl" mkSome :: Maybe (Some (Sum ((:~:) Int) ((:~:) Bool)))
Just (Some (InL Refl))

>>> greadMaybe "garbage" mkSome :: Maybe (Some ((:~:) Int))
Nothing